+++ demo/blocks_textbox_max_lines/run.py
@@ -1,9 +1,7 @@
 import gradio as gr
 
-
 def greet(name: str, repeat: float):
     return "Hello " + name * int(repeat) + "!!"
-
 
 demo = gr.Interface(
     fn=greet, inputs=[gr.Textbox(lines=2, max_lines=4), gr.Number()], outputs="textarea"