+++ demo/stock_forecast/run.py
@@ -3,7 +3,6 @@ import numpy as np
 
 import gradio as gr
 
-
 def plot_forecast(final_year, companies, noise, show_legend, point_style):
     start_year = 2020
     x = np.arange(start_year, final_year + 1)
@@ -19,7 +18,6 @@ def plot_forecast(final_year, companies, noise, show_legend, point_style):
     if show_legend:
         plt.legend(companies)
     return fig
-
 
 demo = gr.Interface(
     plot_forecast,