+++ test/components/test_json.py
@@ -26,8 +26,12 @@ class TestJSON:
             "name": "json",
             "proxy_url": None,
             "_selectable": False,
+            "open": False,
             "key": None,
+            "show_indices": False,
         }
+        js_component = gr.Json(value={"a": 1, "b": 2})
+        assert js_component.get_config()["value"] == {"a": 1, "b": 2}
 
     def test_chatbot_selectable_in_config(self):
         with gr.Blocks() as demo:
@@ -72,9 +76,7 @@ class TestJSON:
             ["F", 30],
         ]
         assert (
-            await iface.process_api(
-                0, [{"data": y_data, "headers": ["gender", "age"]}], state={}
-            )
+            await iface.process_api(0, [{"data": y_data, "headers": ["gender", "age"]}])
         )["data"][0].model_dump() == {
             "M": 35,
             "F": 25,