+++ test/test_chat_interface.py
@@ -43,7 +43,7 @@ def echo_system_prompt_plus_message(message, history, system_prompt, tokens):
 class TestInit:
     def test_no_fn(self):
         with pytest.raises(TypeError):
-            gr.ChatInterface()
+            gr.ChatInterface()  # type: ignore
 
     def test_configuring_buttons(self):
         chatbot = gr.ChatInterface(double, submit_btn=None, retry_btn=None)
@@ -83,6 +83,7 @@ class TestInit:
         chatbot = gr.ChatInterface(double)
         dependencies = chatbot.fns.values()
         textbox = chatbot.textbox._id
+        assert chatbot.submit_btn
         submit_btn = chatbot.submit_btn._id
         assert next(
             (
@@ -92,6 +93,7 @@ class TestInit:
             ),
             None,
         )
+        assert chatbot.retry_btn and chatbot.clear_btn and chatbot.undo_btn
         for btn_id in [
             chatbot.retry_btn._id,
             chatbot.clear_btn._id,
@@ -122,7 +124,9 @@ class TestInit:
             chatbot = gr.ChatInterface(
                 double, examples=["hello", "hi"], cache_examples="lazy"
             )
-            async for _ in chatbot.examples_handler.async_lazy_cache(0, "hello"):
+            async for _ in chatbot.examples_handler.async_lazy_cache(
+                (0, ["hello"]), "hello"
+            ):
                 pass
             prediction_hello = chatbot.examples_handler.load_from_cache(0)
             assert prediction_hello[0].root[0] == ("hello", "hello hello")
@@ -243,6 +247,7 @@ class TestAPI:
     def test_get_api_info(self):
         chatbot = gr.ChatInterface(double)
         api_info = chatbot.get_api_info()
+        assert api_info
         assert len(api_info["named_endpoints"]) == 1
         assert len(api_info["unnamed_endpoints"]) == 0
         assert "/chat" in api_info["named_endpoints"]