+++ test/test_helpers.py
@@ -503,7 +503,7 @@ class TestProcessExamples:
                 [["John"], ["Mary"]],
                 fn=predict,
                 inputs=[t1],
-                outputs=[t2, c],
+                outputs=[t2, c],  # type: ignore
                 cache_examples=True,
             )
 
@@ -594,6 +594,40 @@ class TestProcessExamples:
 
         response = client.post("/api/load_example/", json={"data": [1]})
         assert response.json()["data"] == ["Michael", "Jordan", "Michael Jordan"]
+
+    def test_end_to_end_lazy_cache_examples(self, patched_cache_folder):
+        def image_identity(image, string):
+            return image
+
+        with gr.Blocks() as demo:
+            i1 = gr.Image()
+            t = gr.Textbox()
+            i2 = gr.Image()
+
+            gr.Examples(
+                examples=[
+                    ["test/test_files/cheetah1.jpg", "cheetah"],
+                    ["test/test_files/bus.png", "bus"],
+                ],
+                inputs=[i1, t],
+                outputs=[i2],
+                fn=image_identity,
+                cache_examples="lazy",
+                api_name="load_example",
+            )
+
+        app, _, _ = demo.launch(prevent_thread_lock=True)
+        client = TestClient(app)
+
+        response = client.post("/api/load_example/", json={"data": [0]})
+        data = response.json()["data"]
+        assert data[0]["value"]["path"].endswith("cheetah1.jpg")
+        assert data[1]["value"] == "cheetah"
+
+        response = client.post("/api/load_example/", json={"data": [1]})
+        data = response.json()["data"]
+        assert data[0]["value"]["path"].endswith("bus.png")
+        assert data[1]["value"] == "bus"
 
 
 def test_multiple_file_flagging(tmp_path):
@@ -709,6 +743,7 @@ class TestProgressBar:
 
             button.click(greet, name, greeting)
         demo.queue(max_size=1).launch(prevent_thread_lock=True)
+        assert demo.local_url
 
         client = grc.Client(demo.local_url)
         job = client.submit("Gradio")
@@ -754,6 +789,7 @@ class TestProgressBar:
 
             button.click(greet, name, greeting)
         demo.queue(max_size=1).launch(prevent_thread_lock=True)
+        assert demo.local_url
 
         client = grc.Client(demo.local_url)
         job = client.submit("Gradio")
@@ -795,6 +831,7 @@ class TestProgressBar:
 
         demo = gr.Interface(greet, "text", "text")
         demo.queue().launch(prevent_thread_lock=True)
+        assert demo.local_url
 
         client = grc.Client(demo.local_url)
         job = client.submit("Gradio")
@@ -834,6 +871,7 @@ class TestProgressBar:
 
         demo = gr.Interface(greet, "text", "text")
         demo.queue().launch(prevent_thread_lock=True)
+        assert demo.local_url
 
         client = grc.Client(demo.local_url)
         job = client.submit("Jon")
@@ -880,6 +918,7 @@ async def test_info_isolation(async_handler: bool):
     demo.launch(prevent_thread_lock=True)
 
     async def session_interaction(name, delay=0):
+        assert demo.local_url
         client = grc.Client(demo.local_url)
         job = client.submit(name)