+++ js/colorpicker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/colorpicker
 
+## 0.3.9
+
+### Dependency updates
+
+- @gradio/statustracker@0.7.2
+- @gradio/atoms@0.7.7
+
 ## 0.3.8
 
 ### Dependency updates