+++ js/highlightedtext/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/highlightedtext
 
+## 0.7.3
+
+### Dependency updates
+
+- @gradio/statustracker@0.7.2
+- @gradio/atoms@0.7.7
+
 ## 0.7.2
 
 ### Dependency updates